fix(#48): 홈페이지 개선 /home 라우트 삭제
