--- src/app/home/layout.tsx
@@ -1,13 +1,0 @@
-'use client';
-
-import NavBar from '@/components/common/NavBar';
-
-const HomeLayout = ({ children }: { children: React.ReactNode }) => {
-  return (
-    <div>
-      {children}
-      <NavBar />
-    </div>
-  );
-};
-export default HomeLayout;
