+++ src/app/page.tsx
@@ -1,13 +1,13 @@
 "use client";
 
 import GATracker from '@/components/GATracker';
-import HomePage1 from '@/components/home/HomePage1'
+import HomePage from '@/components/home/HomePage'
 
 export default function Home() {
   return (
     <>
       <GATracker />
-      <HomePage1 />
+      <HomePage />
     </>
   )
 }
